test: split admin stats traces tests
